Adding animations.
Now, in this final part of project I'm adding the animations. I found one problem in the animation of main__heading-books. Theres two animations: The first begins with opacity:0 and ends with opacity:1 and the second vice-versa, but the second one is overlaping the opacity of the first animation. I trying to solve this problem.