The animations are complete
All animations that I wanted to put are finished. There was a problem in the animation of book's page. The second animation, texting-out-contract were overlaping my the first animation. I solved this one put all properties of keyframes in just one animation and setting their percentages. After it, I decided to change the animation.