Solve the position of burger-wrapper and beginning to change photos of slider
I solved the fix of position of burger-menu creating a variable that contains the dimension of my menu. Therefore, my menu always be in center of viewport on mobile resolution. Also, I chose to get another photos with a smaller resolution to set as background of my slider. Miss solve the biggest img's. I'll modify the resolution to solve this problem.